Externalized the internal data
Up to the last commit, all the internal data were produced on the fly
for testing and vignettes every time they were run. This caused long
process time for testing and vignettes. Therefore, these internal data
and models were prepared in prior and saved in ./R/sysdata.rda. The
script that produced this rda file is found in
./data-raw/internal_data.R.

On branch master
Changes to be committed:
      new file:   R/sysdata.rda
      new file:   data-raw/internal_data.R
      modified:   tests/testthat/_snaps/test.md
      modified:   tests/testthat/test.R
      renamed:    vignettes/quick_start.Rmd -> vignettes/partial_sum.Rmd