GAMs fitted and saved as internal data
Fitting GAMs created an issue on CRAN and the package did not go through
the automatic checking on CRAN. This problem was created due to
fitting-process of GAMS, in which multithreading is used. To avoid this
issue, I fitted all the internally-necessary GAMs beforehand and saved
as internal data. The script for that is "data-raw/internal_data.R". The
data created by the script is "R/sysdata.rda".

On branch master
Changes to be committed:
      modified:   R/sysdata.rda
      modified:   cran-comments.md
      modified:   data-raw/internal_data.R
      modified:   tests/testthat/helper-config.R